Fix rep & weight refresh bug
Upon refreshing the page, the order in which data was entered would be presented
inaccurately...things got moved around. This was due to firebase firestore being
a NoSQL database and things were being fetched somewhat randomly. Added a "sets"
field to help render things in proper order. Added some css for the exerciseCard